# feat: implement bart-based trie for superior performance