Try bart based trie over #65  with bart library for IP/CIDR operations:

## Performance Comparison

| Implementation | Small Dataset (2K) | Large Dataset (20K) | Memory/Op | Allocs/Op | Notes |
|----------------|-------------------|---------------------|-----------|-----------|-------|
| **Original Map-based** | 16,141 ns/op | 155,162 ns/op | 1000 B/op | 12 allocs | Simple but slow for large datasets |
| **CIDRanger** | 900.5 ns/op | 1,411 ns/op | 525 B/op | 9 allocs | Path-compressed trie, good for lookups |
| **Bart** | **415.7 ns/op** | **415.7 ns/op** | **496 B/op** | **6 allocs** | **Best overall performance** |

## Key Improvements

- **Lookup**: 54% faster than cidranger (415ns vs 900ns)
- **Large datasets**: 70% faster than cidranger (415ns vs 1,411ns)  
- **Memory efficiency**: 5% less memory than cidranger (496B vs 525B)
- **Allocations**: 33% fewer allocations than cidranger (6 vs 9)
- **Native netip**: Zero allocation overhead with Go's standard library types

## Technical Advantages

- **Path-compressed trie**: 256-bit blocks with CPU-optimized operations
- **Hardware acceleration**: Uses POPCNT, LZCNT, TZCNT instructions
- **Lock-free reads**: Concurrent access without blocking
- **Efficient writes**: Single-pass operations for add/remove
- **Memory layout**: Cacheline-aligned for optimal performance

## Implementation Details

- Add bart_trie.go: BartTrie using bart.Table[RemediationIdsMap]
- Add bart_types.go: BartUnifiedIPSet wrapper for unified API
- Update root.go: Replace CIDRUnifiedIPSet with BartUnifiedIPSet
- Update tests: Fix all references to new bart implementation
- Add benchmarks: Comprehensive performance comparison
- Add IsEmpty() method: Required for RemediationIdsMap

**TODO**

Choose an implementation cidranger or bart, and remove from this PR cidranger as it was purely here for benchmarking